added decimal quantity support

+++ backend/server.js
@@ -313,7 +313,7 @@ app.post('/api/products/:id/transaction', auth, async (req, res) => {
       });
     }
 
-    const qty = parseInt(amount);
+    const qty = parseFloat(amount);
 
     let newQty = product.quantity;
